hilp_zsrl/requirements.txt
numpy
torch
torchvision
hydra-core
hydra-submitit-launcher
scikit-learn
pandas
matplotlib
moviepy
mujoco_py
dm-control==1.0.8
wandb
tensorboard
setuptools>=65.5.1
opencv-python
hiplot
termcolor
d4rl
gym==0.23.1
mujoco==2.3.7